refactor(app): split app.rs into domain modules
app.rs had grown to ~2640 lines (half of it tests). Move-only split
into src/app/, no behaviour change:

- mod.rs — Panel, AppState struct and the core impl (updates, focus,
  toasts, refresh deadline) + re-exports keeping crate::app::* paths
  stable for input.rs / ui / main.rs.
- inputs.rs — all wizard input enums + RESOURCE_*/DETACH_MODES.
- scan.rs — ScanFilter, scanner objects/actions types and every
  scan-history/scanner method.
- travel.rs, map.rs, waypoints.rs, message.rs — per-domain.
- inventory.rs — InventoryRow, rows/cursor, jettison builder, recipes.
- mannies.rs — selection, collect_* candidates and the manny-wizard
  input/error setters.
- tests.rs — the existing unit tests, unchanged, in one #[cfg(test)]
  module file (a per-domain split of tests can come later).

clamp_inventory_selection and probe_current_sector_scan become
pub(crate) (called across sibling modules).
