+++ src/app/message.rs
@@ -1,0 +1,38 @@
+use crate::api::types::{CraftingRecipe, Manny, Probe, ProbeInventory, ProbeMovement, SectorObservation, VisitedSector};
+
+pub enum ApiMessage {
+    ProbeUpdated(Probe),
+    ManniesUpdated(Vec<Manny>),
+    SectorUpdated(SectorObservation),
+    ScanError(String),
+    MoveStarted(ProbeMovement),
+    MoveError(String),
+    RepairStarted,
+    RepairError(String),
+    MineStarted,
+    MineError(String),
+    VersionFetched(u32),
+    VisitedSectorsFetched(Vec<VisitedSector>),
+    JettisonDone(ProbeInventory),
+    JettisonError(String),
+    CraftStarted,
+    CraftError(String),
+    SalvageStarted,
+    SalvageError(String),
+    RecallStarted,
+    RecallError(String),
+    DeployStarted,
+    DeployError(String),
+    AtomicPrinterCraftStarted,
+    AtomicPrinterCraftError(String),
+    RecipesFetched(Vec<CraftingRecipe>),
+    RenameMannyDone(Manny),
+    RenameMannyError(String),
+    InspectStarted,
+    InspectError(String),
+    RecoverStarted,
+    RecoverError(String),
+    DetachStarted,
+    DetachError(String),
+    Error(String),
+}
